Play user input
